requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.4.0
psycopg2-binary==2.9.10 # only if you later use Postgres; SQLite needs nothing
channels==4.1.0
daphne==4.1.2
